refactor(OutputView): ProgressView, ResultView로 분리
