+++ src/main/kotlin/omok/Application.kt
@@ -2,9 +2,15 @@ package omok
 
 import omok.controller.OmokController
 import omok.view.InputView
-import omok.view.OutputView
+import omok.view.ProgressView
+import omok.view.ResultView
 
 fun main() {
-    val controller = OmokController(InputView(), OutputView())
+    val controller =
+        OmokController(
+            InputView(),
+            ProgressView(),
+            ResultView(),
+        )
     controller.start()
 }
