+++ src/main/kotlin/omok/controller/OmokController.kt
@@ -5,22 +5,27 @@ import omok.model.board.Board
 import omok.model.stone.BlackStone
 import omok.model.stone.WhiteStone
 import omok.view.InputView
-import omok.view.OutputView
+import omok.view.ProgressView
+import omok.view.ResultView
 
-class OmokController(private val inputView: InputView, private val outputView: OutputView) {
+class OmokController(
+    private val inputView: InputView,
+    private val progressView: ProgressView,
+    private val resultView: ResultView,
+) {
     fun start() {
         val omokGame = readyOmokGame()
 
         omokGame.start(
             { inputView.readPosition(it) },
-            { outputView.drawBoard(it) },
-            { outputView.printWinner(it) },
+            { progressView.drawBoard(it) },
+            { resultView.printWinner(it) },
         )
     }
 
     private fun readyOmokGame(): OmokGame {
-        outputView.printStartGameComment()
-        outputView.drawBoard(Board)
+        progressView.printStartGameComment()
+        progressView.drawBoard(Board)
         return OmokGame(BlackStone(), WhiteStone())
     }
 }
