+++ src/main/kotlin/omok/view/ProgressView.kt
@@ -1,10 +1,9 @@
 package omok.view
 
 import omok.model.board.Board
-import omok.model.stone.GoStone
 import omok.model.stone.Stone
 
-class OutputView {
+class ProgressView {
     fun printStartGameComment() = println(START_GAME_MESSAGE)
 
     fun drawBoard(board: Board) {
@@ -17,15 +16,6 @@ class OutputView {
             printBoard(board, row)
         }
         printRowValue()
-    }
-
-    fun printWinner(stone: GoStone) {
-        lineBreak()
-        if (stone.stoneType == Stone.BLACK_STONE) {
-            println(BLACK_STONE_WIN_MESSAGE)
-        } else {
-            println(WHITE_STONE_WIN_MESSAGE)
-        }
     }
 
     private fun printBoard(
@@ -103,8 +93,6 @@ class OutputView {
             Stone.NONE -> ""
         }
 
-    private fun lineBreak() = println()
-
     companion object {
         private const val UPPER_LENGTH = 14
         private const val LOWER_ROW = 0
@@ -113,7 +101,5 @@ class OutputView {
         private const val MIN_COL = 0
         private val ROW_RANGE = ROW_START..ROW_END
         private const val START_GAME_MESSAGE = "오목 게임을 시작합니다."
-        private const val BLACK_STONE_WIN_MESSAGE = "흑의 승리입니다."
-        private const val WHITE_STONE_WIN_MESSAGE = "백의 승리입니다."
     }
 }
